+++ test/runtests.jl
@@ -72,6 +72,7 @@ end
         include("unit/test_strain.jl")
         include("unit/test_magnetoelastic.jl")
         include("unit/test_magnonphonon.jl")
+        include("unit/test_strainedmodels.jl")
         include("unit/test_latticeonly.jl")
         include("unit/test_effective.jl")
         include("unit/test_threading.jl")